requirements.txt
# Core dependencies
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
pyobjc-core>=9.0
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0

# Testing dependencies
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-mock>=3.12.0
hypothesis>=6.92.0

# Development dependencies
coverage>=7.3.0

# Build dependencies
py2app>=0.28.0
